requirements.txt
openai
python-dotenv
streamlit
streamlit-shadcn-ui
pycryptodome
matplotlib
pandas
scikit-learn
requests